[BluetoothAudioSink] Extract AudioPlayer to separate file
